LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2-expected.txt

Button1  Button2  Invoker1  Button3  Button4
Invoker  after
Show popover Hide popover

FAIL Popover focus navigation assert_equals: Focus should move from invoker into the open popover expected Element node <button id="inside_popover1">Inside1</button> but got Element node <button id="button3">Button3</button>
PASS Circular reference tab navigation
FAIL Popover focus returns when popover is hidden by invoker assert_equals: Hide button should be focused due to autofocus attribute expected Element node <button popovertarget="deleted1" popovertargetaction="hid... but got Element node <button popovertarget="deleted1" popovertargetaction="sho...
